Moves EventCard component to events folder
