+++ components/events/EventCard.vue
